requirements.txt
Flask>=2.0.0
Flask-Login>=0.5.0
Flask-WTF>=0.14.3
Flask-migrate>=3.0.0
Flask-Mail>=0.9.1
Flask-SQLAlchemy>=2.5.1
mysql-connector-python>=8.0.23
gunicorn>=20.1.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-cov>=4.0.0
requests>=2.25.1
email-validator>=2.0.0
WTForms>=3.0.0